refactor dpm wave item routes
