+++ src/app/router_registry.py
@@ -95,6 +95,7 @@ from app.routers.dpm_wave_campaign_workflow_boards import (
     router as dpm_wave_campaign_workflow_boards_router,
 )
 from app.routers.dpm_wave_evidence import router as dpm_wave_evidence_router
+from app.routers.dpm_wave_items import router as dpm_wave_items_router
 from app.routers.dpm_waves import router as dpm_waves_router
 from app.routers.foundation import router as foundation_router
 from app.routers.intake import router as intake_router
@@ -214,6 +215,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_construction_router)
     app.include_router(dpm_proof_packs_router)
     app.include_router(dpm_waves_router)
+    app.include_router(dpm_wave_items_router)
     app.include_router(dpm_wave_actions_router)
     app.include_router(dpm_wave_evidence_router)
     app.include_router(dpm_wave_ai_router)
