+++ src/app/routers/dpm_wave_actions.py
@@ -1,50 +1,14 @@
 from fastapi import APIRouter, Path
 
-from app.contracts.dpm_waves import (
-    DpmWaveErrorDetail,
-    DpmWaveForwardRequest,
-    DpmWaveGatewayResponse,
-)
+from app.contracts.dpm_waves import DpmWaveForwardRequest, DpmWaveGatewayResponse
 from app.middleware.correlation import correlation_id_var
-from app.routers.dpm_openapi import manage_upstream_error_responses
+from app.routers.dpm_wave_action_common import UPSTREAM_WAVE_ACTION_ERROR_RESPONSES
 from app.services.dpm_service_provider import dpm_wave_service
 
 router = APIRouter(
     prefix="/api/v1/dpm/command-center/waves",
     tags=["DPM Command Center"],
 )
-_UPSTREAM_ERROR_RESPONSES = manage_upstream_error_responses(
-    error_model=DpmWaveErrorDetail,
-    not_found_description="lotus-manage could not find the requested rebalance wave action.",
-    conflict_description="lotus-manage rejected the rebalance-wave action as conflicting.",
-    invalid_payload_description=(
-        "lotus-manage rejected the rebalance-wave action payload as invalid."
-    ),
-    unavailable_description=(
-        "lotus-manage rebalance-wave action authority is unavailable or degraded."
-    ),
-)
-
-
-@router.get(
-    "/{wave_id}/items",
-    response_model=DpmWaveGatewayResponse,
-    summary="List DPM rebalance wave items",
-    description=(
-        "What: returns manage-owned item-level wave posture. When: call this for Workbench item "
-        "tables, source-readiness review, construction selection, proof-pack linkage, and "
-        "handoff readiness. How: Gateway preserves item states, reason codes, diagnostics, refs, "
-        "and aggregate metrics without deriving readiness."
-    ),
-    responses=_UPSTREAM_ERROR_RESPONSES,
-)
-async def get_wave_items(
-    wave_id: str = Path(..., description="Manage-owned rebalance-wave identifier."),
-) -> DpmWaveGatewayResponse:
-    return await dpm_wave_service().get_wave_items(
-        wave_id=wave_id,
-        correlation_id=correlation_id_var.get(),
-    )
 
 
 @router.post(
@@ -57,7 +21,7 @@ async def get_wave_items(
         "Gateway forwards controls unchanged and preserves manage item classifications and "
         "supportability; it never promotes caller-supplied portfolio ids to ready."
     ),
-    responses=_UPSTREAM_ERROR_RESPONSES,
+    responses=UPSTREAM_WAVE_ACTION_ERROR_RESPONSES,
 )
 async def source_check_wave(
     request: DpmWaveForwardRequest,
@@ -80,7 +44,7 @@ async def source_check_wave(
         "preserves manage construction refs, item states, and degradation reasons without "
         "building holdings, market data, model targets, or alternatives locally."
     ),
-    responses=_UPSTREAM_ERROR_RESPONSES,
+    responses=UPSTREAM_WAVE_ACTION_ERROR_RESPONSES,
 )
 async def simulate_wave(
     request: DpmWaveForwardRequest,
@@ -88,31 +52,6 @@ async def simulate_wave(
 ) -> DpmWaveGatewayResponse:
     return await dpm_wave_service().simulate_wave(
         wave_id=wave_id,
-        body=request.body,
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.post(
-    "/{wave_id}/items/{wave_item_id}/select",
-    response_model=DpmWaveGatewayResponse,
-    summary="Select DPM wave item alternative",
-    description=(
-        "What: records a manage-owned construction alternative selection for one wave item. "
-        "When: call this after PM/CIO review of generated alternatives. How: Gateway forwards "
-        "selection, actor, reason, comment, and proof-pack-generation controls unchanged and "
-        "preserves manage selection, proof-pack, and degraded posture."
-    ),
-    responses=_UPSTREAM_ERROR_RESPONSES,
-)
-async def select_wave_item(
-    request: DpmWaveForwardRequest,
-    wave_id: str = Path(..., description="Manage-owned rebalance-wave identifier."),
-    wave_item_id: str = Path(..., description="Manage-owned rebalance-wave item identifier."),
-) -> DpmWaveGatewayResponse:
-    return await dpm_wave_service().select_wave_item(
-        wave_id=wave_id,
-        wave_item_id=wave_item_id,
         body=request.body,
         correlation_id=correlation_id_var.get(),
     )
@@ -128,7 +67,7 @@ async def select_wave_item(
         "manage approval state and exceptions without approving blocked, degraded, or unselected "
         "items locally."
     ),
-    responses=_UPSTREAM_ERROR_RESPONSES,
+    responses=UPSTREAM_WAVE_ACTION_ERROR_RESPONSES,
 )
 async def approve_wave(
     request: DpmWaveForwardRequest,
@@ -150,7 +89,7 @@ async def approve_wave(
         "internal operations handoff. How: Gateway preserves manage staged state and exceptions "
         "without treating staging as external execution."
     ),
-    responses=_UPSTREAM_ERROR_RESPONSES,
+    responses=UPSTREAM_WAVE_ACTION_ERROR_RESPONSES,
 )
 async def stage_wave(
     request: DpmWaveForwardRequest,
@@ -173,7 +112,7 @@ async def stage_wave(
         "refs and the `external_execution_claimed=false` boundary; it does not send orders or "
         "claim client/execution completion."
     ),
-    responses=_UPSTREAM_ERROR_RESPONSES,
+    responses=UPSTREAM_WAVE_ACTION_ERROR_RESPONSES,
 )
 async def handoff_wave(
     request: DpmWaveForwardRequest,
@@ -196,7 +135,7 @@ async def handoff_wave(
         "cancellation diagnostics and does not cancel external orders because RFC-0041 handoff is "
         "internal readiness evidence only."
     ),
-    responses=_UPSTREAM_ERROR_RESPONSES,
+    responses=UPSTREAM_WAVE_ACTION_ERROR_RESPONSES,
 )
 async def cancel_wave(
     request: DpmWaveForwardRequest,
